# backend/openapi_doc/specs/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Routine Cloud API v1",
    "description": "Routine Cloud API SDK",
    "version": "v1"
  },
  "paths": {
    "/devices/": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "List Devices",
        "description": "Return all devices that belong to the authenticated user.",
        "operationId": "devices_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceRead"
                  },
                  "title": "Response Devices List"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Get Device",
        "description": "Return a single device by ID if it belongs to the authenticated user.",
        "operationId": "devices_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/friends/": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "List Friends",
        "description": "List all accepted friends of the current user.",
        "operationId": "friends_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendRead"
                  },
                  "title": "Response Friends List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/friends/requests": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "List Friend Requests",
        "description": "List all pending friend requests sent TO the current user.",
        "operationId": "friends_requests_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendRead"
                  },
                  "title": "Response Friends Requests List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/friends/add": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Add Friend",
        "description": "Send a friend request using a friend code.",
        "operationId": "friends_add",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/friends/{friend_id}/accept": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Accept Friend",
        "description": "Accept a friend request.",
        "operationId": "friends_accept",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Friend Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/friends/{friend_id}/decline": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Decline Friend",
        "description": "Decline a friend request.",
        "operationId": "friends_decline",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Friend Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/friends/{friend_id}": {
      "delete": {
        "tags": [
          "friends"
        ],
        "summary": "Remove Friend",
        "description": "Remove a friend or cancel a request.",
        "operationId": "friends_remove",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Friend Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeviceRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus",
            "default": "offline"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "id"
        ],
        "title": "Device"
      },
      "DeviceStatus": {
        "type": "string",
        "enum": [
          "online",
          "offline"
        ],
        "title": "DeviceStatus"
      },
      "FriendAdd": {
        "properties": {
          "friend_code": {
            "type": "string",
            "title": "Friend Code"
          }
        },
        "type": "object",
        "required": [
          "friend_code"
        ],
        "title": "FriendAdd"
      },
      "FriendRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "profile_picture": {
            "title": "Profile Picture",
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/FriendshipStatus",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "FriendRead"
      },
      "FriendshipStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted"
        ],
        "title": "FriendshipStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}